refactor(core): ♻️ move all arithmetic logic to dedicated arithmetic module
